feat(bivariate-lengends): color legend view refactoring
