+++ src/views/BivariateManager/sub_views/ColorLegends/ColorLegendsView.tsx
@@ -1,7 +1,4 @@
-import { Select } from '@konturio/ui-kit';
 import { useAtom } from '@reatom/react';
-import { useMemo } from 'react';
-import { i18n } from '~core/localization';
 import { KonturSpinner } from '~components/LoadingSpinner/KonturSpinner';
 import { SentimentsCombinationsList } from '~features/bivariate_color_manager/components';
 import { bivariateColorManagerResourceAtom } from '~features/bivariate_color_manager/atoms/bivariateColorManagerResource';
@@ -11,17 +8,11 @@ import {
 } from '~features/bivariate_color_manager/components/CssTransitionWrapper/CssTransitionWrapper';
 import { LegendWithMap } from '~features/bivariate_color_manager/components/LegendWithMap/LegendWithMap';
 import { bivariateColorManagerAtom } from '~features/bivariate_color_manager/atoms/bivariateColorManager';
+import { ColorLegendFilters } from '~features/bivariate_color_manager/components/ColorLegendFilters/ColorLegendFilters';
 import style from './ColorLegendsView.module.css';
 import type { LayerSelectionInput } from '~features/bivariate_color_manager/atoms/bivariateColorManager';
 import type { LayerSelectionFull } from '~features/bivariate_color_manager/components/LegendWithMap/LegendWithMap';
 import type { BivariateColorManagerData } from '~features/bivariate_color_manager/atoms/bivariateColorManagerResource';
-import type { SelectItemType } from '@konturio/ui-kit/tslib/Select/types';
-
-const LayersFilterMenuClasses = { menu: style.LayersFilterMenu };
-
-function itemToString(item) {
-  return item ? item.title : '';
-}
 
 function isFullSelection(
   selection: LayerSelectionInput | null,
@@ -33,8 +24,8 @@ function isFullSelection(
 
 export const ColorLegendsView = () => {
   const [
-    { filteredData: data, indicators, layersSelection, selectedRows, filters },
-    { setLayersSelection, setLayersFilter, setSelectedRows },
+    { filteredData: data, layersSelection, selectedRows, filters },
+    { setLayersSelection, setSelectedRows },
   ] = useAtom(bivariateColorManagerAtom);
   const [{ loading }] = useAtom(bivariateColorManagerResourceAtom);
   const filteredDataNotEmpty = data && Object.keys(data).length > 0;
@@ -44,29 +35,11 @@ export const ColorLegendsView = () => {
       : null;
   const anyFilterActivated = Object.values(filters).filter(Boolean).length > 0;
 
-  const selectIndicatorsData: SelectItemType[] = useMemo(() => {
-    return (
-      indicators?.map((indicator) => ({
-        title: indicator.label,
-        value: indicator.name,
-      })) || []
-    );
-  }, [indicators]);
-
   return (
     <>
       <div className={style.List}>
-        <div className={style.ListFilters}>
-          <Select
-            onChange={(item) => setLayersFilter(item.selectedItem || null)}
-            classes={LayersFilterMenuClasses}
-            items={selectIndicatorsData}
-            itemToString={itemToString}
-            disabled={loading}
-          >
-            {i18n.t('bivariate.color_manager.layers_filter')}
-          </Select>
-        </div>
+        <ColorLegendFilters />
+
         <div className={style.ListBody}>
           {loading ? (
             <KonturSpinner />
